Use local and ip address to create vxlan interface
Linux kernel does not allow a loopback interface to be an endpoint dev interface for VXLAN terminations. In order to allow loopback interfaces to be VXLAN termination endpoints, we need to change how the vxlan is created: don't use dev, use local and an ip address. Using local is also, perhaps, a more flexible solution, since it allows for picking the specific IP on the interface, as opposed to whatever the OS decides to use.